backend/requirements.txt
# ========================================
# KTV 报表系统 - Python 依赖
# ========================================

# Web 框架
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# 数据库
sqlalchemy==2.0.25
pymysql==1.1.0
alembic==1.13.1

# 数据处理
pandas==2.1.4
openpyxl==3.1.2
xlrd==2.0.1
numpy==1.26.3
chardet==5.2.0

# 数据验证
pydantic==2.10.6
pydantic-settings==2.1.0
email-validator==2.1.0

# 认证
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# 工具
python-dotenv==1.0.0
httpx==0.26.0

# 开发工具
pytest==7.4.4
pytest-asyncio==0.23.3
black==23.12.1
flake8==7.0.0

